requirements.txt
django==1.9.2
django-GNU-Terry-Pratchett==0.2.0
mysqlclient==1.3.7
django-markup==1.0 # TODO: needed?
markdown==2.6.5    # TODO: needed?
django-zurb-foundation-rtl==6.0.6.3
django-pagedown==0.1.1
